refactor : MemberDTO 이동으로 인한 변경
