+++ src/main/java/gift/service/KakaoService.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.config.KakaoProperties;
 import gift.dto.KakaoAccessTokenDTO;
 import gift.dto.KakaoUserInfoDTO;
-import gift.dto.MemberDTO;
+import gift.dto.memberDTO.MemberDTO;
 import gift.dto.OrderResponseDTO;
 import gift.dto.TemplateObjectDTO;
 import gift.model.Member;
