+++ src/main/java/gift/service/MemberService.java
@@ -1,6 +1,6 @@
 package gift.service;
 
-import gift.dto.MemberDTO;
+import gift.dto.memberDTO.MemberDTO;
 import gift.model.Member;
 import gift.repository.MemberRepository;
 import org.springframework.stereotype.Service;
